Don't send up-to-date check telemetry during validation
When a project is configured to validate for incremental build failures, we run the fast up-to-date check a second time after the build. If the project is not up-to-date, then there is an incremental build failure.

For that second run, we should not send the regular check's telemetry. We have different telemetry to track the failure.