Rework WarningLevel property
Historically, the `WarningLevel` property (corresponding to the `-warn:<n>` switch) controlled what set of warnings would be produced by the C# compiler. "0" meant produce no warnings (only errors). "1" meant produce only the most severe warnings. "2" meant produce all the warnings in 1, plus a set of less severe warnings. "3" meant all the warnings in 2 (and 1) plus additional, even less severe warnings. And so on. The default for a long time was "4" which effectively turned on all warnings the compiler could produce. The general idea was that, if need be, the user could pick a low `WarningLevel` to identify and subsequently fix the most impactful warnings in their code, and gradually ratchet the level up over time.

The scope and effective behavior of `WarningLevel` changed with .NET 5 and C# 9.0. This added a new warning level, 5, which controlled not only additional warnings added with C# 9.0 support, but _also_ certain analyzers included with the .NET 5.0 SDK. The SDK also sets the `WarningLevel` to 5 by default. Similarly, the .NET 6.0 SDK adds level 6, sets that as the default, and adds new analyzers/warnings tied to that level. The idea now is that as you adopt newer versions of the SDK you get new warnings, but have the option of turning them off by reducing the `WarningLevel`.

Along the way the logic of how `WarningLevel` is set to a default value in the SDK has become complicated. If you are targeting .NET 5.0 or later, the `WarningLevel` is computed based on the `EffectiveAnalysisLevel` property, which is itself computed from `AnalysisLevel`. If the `EffectiveAnalysisLevel` is 5.0 or greater, the SDK targets will override any `WarningLevel` set in the project in favor of specific values that correspond to the `EffectiveAnalysisLevel`. In these cases it doesn't make sense to provide controls to the user to change `WarningLevel`; the new value will be written to the project file but won't be honored, and the control will go back to the value provided by the SDK.

Here we address this by adding a hidden, synthetic property named `WarningLevelOverridden`. When it is true, we replace the `WarningLevel` controls with an equivalent read-only version and point the user to the `AnalysisLevel` property. A property interceptor computes the value of `WarningLevelOverridden` based on `EffectiveAnalysisLevel`.