Up-to-date check tracks build start time
The fast up-to-date check cannot rely solely on file timestamps to identify when a build is required. We also need to know when a build has occurred following a change in the project's items.

Historically we tracked time at which the project's set of inputs changed against the time of the last build output. This allowed an overbuild performance bug to occur, where the change in the set of items would not correspond with a change in build outputs. In this state, the project would build on every invocation.

We recently fixed an issue where we would schedule a build after a rebuild occurred. We did that by adding build event tracking, and flagging when a rebuild occurred.

This commit extends the build event tracking so that we know when the last successful build started.

With that information, we no longer check the items changed time against the build output time. Instead, we compare against the last successful build's start time. In this way, we trust that MSBuild has brought the project up-to-date with respect the the change of inputs, and break out of the overbuild loop we would previously have found ourselves in.