Add explicit "General" category
Related to AB#1495090.

In a `Rule`, any property that does not explicitly state which category it belongs to is added to the "General" category, and if that category is not declared in the `Rule` it is generated. The problem with allowing it to be generated is that we don't get a localized `DisplayName` for the category, and the property page UI is left to use the unlocalized name which will always be "General".

Here we add an explicit "General" category to the Debug, Resources, and Settings property pages. We also explicitly add the properties of those pages to the category. I've also gone ahead and copied the translation of the term "General" from the Application property page, as it will be the same.