Harden against project capability changes
`DependencyRulesSubscriber` tracks the set of active `IDependenciesRuleHandler` objects via an `OrderPrecedenceImportCollection`. This collection can change over time in response to dynamic project capability changes.

The previous implementation would fail if the set changed over time. For example, if a new handler was added after initialization, its rule would not be located in the subscription data, leading to a `KeyNotFoundException`.

This commit captures the set of handlers during initialization, and re-uses that collection thereafter. This prevents issues due to exports changing dynamically during the lifetime of the project.