Update fault event level, when DisplayMissingCompnents failed.
The reason behind it is that the task failed in our future VS prototype branch, as VS setup component is not yet working.  Reporting the project has limited service will put the project into a state where several import features are disabled (like operation progress).  This should not have any impact on normal function of this code.