Fast up-to-date check handles changed items sooner
In #8005 we changed how the fast up-to-date check handled changes in the set of project items. That change means we no longer check against build output times.

As a consequence of that, we can do this check sooner, before we scan outputs. In the case that this check fails, it will fail a little faster. It also has the benefit that we only run the check once in the case that we have multiple sets of inputs/outputs, as previously we performed the check for each set.