Express always-hidden property with static data
This property should never be visible. This was being achieved with a visibility condition. However that requires the UI to create and maintain an object for the property, re-evaluating its visibility as the UI updates.

By setting the `IsVisible` property to false, the UI can be slightly more efficient here.