Prevent unnecessary LaunchProfile cloning
Previously code would make defensive copies of `ILaunchProfile` instances, by passing them to the constructor as `new LaunchProfile(launchProfile)`. This guarantees an allocation.

The new `LaunchProfile.Clone` method returns the profile unchanged if it has concrete type `LaunchProfile` (via runtime type test). If so, we know it is immutable so can return it as-is.

If not, we allocate a clone.