Remove redundant Task aliases
Historically one of the MS.VS namespaces had its own `Task` class, which conflicted with `System.Threading.Tasks.Task` in code.

Whenever this conflict occurred, we had to add a `using` alias in the code.

With the change to 17.0 the `Task` type was removed from the VS SDK. These aliases are no longer needed.