Avoid integer overflow converting IntPtr to Int32
Fixes https://dev.azure.com/devdiv/DevDiv/_workitems/edit/1414398

This bug was introduced during the port of VS to 64-bit, where IntPtr could potentially have a 64-bit value. This code used `CType` to convert that value to `Int32`. At runtime, if the value is outside the domain of `Int32` values, an `OverflowException` is thrown. Watson shows a moderate number of crashes related to this.

While we could fix the overflow itself by truncating higher bits, it's likely that any value outside the range of `Int32` is not a valid screen coordinate, or ever would be. If the value is outside that range, use the current mouse cursor position instead.