Stub out a C# build page
The existing BuildPropertyPage.xaml was designed primarily with C# in mind, and as such many of the properties are also applicable to VB and F#, but many are not. We can see this in how it is extended by BuildPropertyPage.FSharp.xaml with most properties being adopted as-is by F#, and half a dozen or so being hidden because they don't apply to that language.

Now that we're starting on VB support for the new property pages it makes sense to keep the properties that are truly common in BuildPropertyPage.xaml, and move the language-specific ones to language-specific pages. This commit takes the first step, stubbing out a BuildPropertyPage.CSharp.xaml that extends BuildPropertyPage.xaml, and (currently) makes no modifications to it.