Add VBDisableAllWarningsValueProvider
The VB "Compile" page has a "Disable all warnings" check box. However, this doesn't correspond to a property in the project file with a true or false value. Rather, if the "WarningLevel" property is 0, all warnings are disabled; if it has any other value, warnings are enabled. The `VBDisableAllWarningsValueProvider` serves to map between the two.