Prohibit sharing of certain IWorkspaceContextHandler implementations
In #8181 we moved these handlers from `ConfiguredProject` scope to `UnconfiguredProject` scope in preparation for further work on #7541.

However five of our handler implementations have per-context state. `ApplyChangesToWorkspaceContext` (in `ConfiguredProject` scope) handles this correctly by using an `ExportFactory` to create private (non-shared) instances of the handlers for each configuration. However that left these exports advertising themselves as safe for use in `UnconfiguredProject` scope, which is only true if they are not shared.

This commit annotates the exports to prevent accidental sharing in future.