Introduce read-only variants of the warning properties
There are a few warning sets that are forced to "Error" when "Option strict" is on, and we currently swap in a read-only variant of the control when that happens.

We also have the option of disabling all warnings, in which case the individual warning sets are forced to "None". Here we create read-only variants of these as well, with visibility conditioned on the WarningSeverity property.