Port AppDotXamlDocument
Port AppDotXamlDocument.vb in Microsoft.VisualStudio.Editors to Microsoft.VisualStudio.ProjectSystem.Managed.VS. In the old property pages for WPF projects, this type supports getting and setting the "Startup URI" and "Shutdown Mode" properties by reading and writing to the Application.xaml via an `IVsTextLines` instance. We will need the same functionality in the new property pages, but also the option to make changes that are not backwards compatible with the old pages. So rather than trying to understand and re-create the behavior in new code I've opted for a straight port of the existing code to C#.

I've opted to move it to C# with only minor changes so we can get the WPF properties working end-to-end as soon as possible, and then we have the option of iterating on the implemenation. So it's not pretty, and not how we would write it now, but we can refactor it later.

As yet nothing consumes this code.