Fix overbuild on project load
Changes in 17.2 intended to fix an overbuild loop inadvertantly introduced a single occurrence of overbuild for the first build of a project, in some cases.

Those changes allow the fast up-to-date check to track the time at which the last successful build started, and use that timestamp in comparisons for other kinds of changes to know whether those changes should trigger subsequent builds.

The issue arose because the "last successful build start time" value is not persisted between sessions. On reloading a project, the value is DateTime.MinValue, which is always before the last set of item changes, and therefore triggers a build. That build then updates the in-memory "last successful build start time", and from that point on the project behaves as desired.

Changes in this commit:

- Persist the "last successful build time" value along with the other state we keep.
- Bump version of `.futdcache` file, as we now store additional information.
- Make `DateTime` values nullable to it's clearer that we need to handle the missing-value case, and remove use of `DateTime.MinValue`.
- Remove redundant `WasStateRestored` property from our snapshot. It's enough to test for a null timestamp in the place that property was read.
- Move handling of build events off the main thread, on to the thread pool.
- Update and extend documentation in code.