Optimise checking of copy markers in the fast up-to-date check
The fast up-to-date check compares a single output marker timestamp against potentially many input marker timestamps.

Previously we checked all inputs markers and remembered the latest one, then compared that with the output marker timestamp.

This change allows this loop to exit early, as soon as an input marker is found that's newer than the output marker.