Improve exception message for users when SettingsDesigner fails to open a file.
This change keeps the original exception message and adds more verbose
describing a high level description of the root cause and a help message
for cases when the file is not a Settings Designer.

This help message is a localized string that suggest to open the file with
the command 'OpenWith' using a text editor.