Add Rule file for PackageVersion items
Related to AB#1544123.

When CPS adds or modifies PackageVersion items, we would like it to persist the "Version" metadata as an attibute rather than as a child element. This can be controlled by providing a `Rule` for PackageVersion items that specifies `PersistenceStyle="Attribute"` for the relevant properties. If the `Rule` is placed in the "File" or "BrowseObject" context then the CPS types responsible for writing item metadata will pick it up and use the appropriate setting.