Only access App.xaml in WPF apps
Fixes AB#1553442.

In #8220 we added VB property page support for WPF-related properties, like "Startup URI" and "Shutdown Mode". Ultimately these properties are read from the Application.xaml file--which generally only exists in WPF projects. The properties themselves are always present in the .xaml file, which means we end up trying to retrieve their values whenever the .xaml file is in use--we just hide the controls unless the project is a WPF application. Unfortunately the current interceptor for this property always tries to access the Application.xaml file, which will fail with an exception if the file isn't present.

To fix this, the interceptor first checks if this is a WPF application before trying to get the values from the Application.xaml. Specifically, we check if "UseWPF" is true, and that the output type of the project is "WinExe".

This also adds a number of unit tests to enforce that the Application.xaml is only accessed at appropriate times.