Optimise language service during configuration switches
This change alters how the language service constructs and manages Roslyn workspace objects.

Previously, a workspace would be created per ConfiguredProject. With this change, one is created per "slice".

The easiest way to explain slices is with an example. Consider the following configurations:

- Debug|AnyCPU|net48
- Debug|AnyCPU|net60
- Release|AnyCPU|net48
- Release|AnyCPU|net60

If "Debug|AnyCPU" is active in the IDE, then the first two configurations are active. If the user switches to "Release", those two are deactivated, and the last two become active.

When considering slices, the configuration (eg. Debug/Release) and platform (eg. AnyCPU) are ignored. Only additional dimensions are considered (eg. TargetFramework).

CPS provides a new facility that allows subscribing to data for (in this example) "net48" and "net60" slices of the project. If the user switches configuration, the same subscription is used, however any difference in the data between configurations is received (via Dataflow) as a delta.

Note that for single-targeting projects (i.e. a single target framework) there is only one slice for the project, and it still updates as described.

Given we now receive a delta when the configuration changes, we are able to pass that delta on to Roslyn. Roslyn will be able to optimise this process to avoid discarding all work done for one configuration, and starting fresh with another configuration when a switch occurs.

We don't have MEF scopes for slices, and slices do not belong to a specific configuration, so MEF components that work with slices needed to be moved from ConfiguredProject scope to UnconfiguredProject scope.

This change also reduces the number of types/interfaces/concepts involved in creating, updating and generally working with workspaces.