All build data to arrive before evaluation data
The previous code assumed evaluation data would always arrive before build data, which is not the case. Because of that assumption, an exception could be seen.

This change reworks the processing of data in the `Workspace` to permit build data to arrive first. We handle this by using two `TaskCompletionSource` objects to signal the integration of both initial evaluation and build data.

We also:

- Pass active configured project for build data too, to allow initialization
- Remove redundant JTF indirection
- TIghten the states and transitions within the Workspace
- Document a bunch of code
- Ensure initialization tasks run asynchronously when signalled
- Overload the `TaskCompletionSource` constructor to accept `TaskCreationOptions`