Avoid deadlock between CodeModel API and build data
Fixes AB#1600170

When project system components outside the language service host want to write to the Roslyn project context, we block those operations until the `Workspace` object is initialised with both evaluation and build data.

This state is signalled via a `TaskCompletionSource`, unblocked by the arrival of the first build data for that project slice.

Previously we used `JTF.RunAsync` in an attempt to join work between components that acquire a lock over the `Workspace`. That does not correctly join the two operations. What we need to do is temporarily join the `JoinableTaskCollection` used within the dataflow handling logic. This allows dataflow to keep pumping on the main thread, even if other code is waiting for a write lock.