Don't delay workspace writers as long
In the prior language service host implementation (replaced a few weeks ago) attempts to lock the IWorkspaceProjectContext would block only on the creation of the context object, not on its population with build data.

Build data can arrive quite a bit later. We do not need to wait that long.

This change moves the signal that triggers this condition earlier, matching the previous behavior.

This fixes a potential UI delay for code paths that access certain APIs such as CodeModel or TempPE.